pyproject.toml
[tool.poetry]
name = "wss-stock"
version = "0.1.0"
description = ""
authors = ["AkimGomes <ak-gomes@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.2.0"
setuptools = "^72.1.0"
pytest-django = "^4.8.0"
drf-yasg = "^1.21.7"
bson = "^0.5.10"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
